add db update script
git-svn-id: http://mvh-source.googlecode.com/svn/trunk/pico@302 f27d9930-b0da-11de-a90c-f79689ea6043

+++ Library/Pico/AdminController.php
@@ -29,84 +29,88 @@ class Pico_AdminController extends Nano_Controller{
     protected  function editAction(){
         $request    = $this->getRequest();
         $klass      = sprintf( 'Model_%s', ucfirst($request->controller) );
-        
+
         $item = $klass::get( $request->id );
-        
+
         if( $request->id == null ){
             // set up defaults for new items
             $count = $item->all()->count();
             $item->name     = sprintf( 'New %s %d', $request->controller, $count);
             $item->visible  = true;
         }
-        
+
         $form = new Form_EditItem( $item );
-        
+
         if( $request->isPost() && $post = $request->getPost() ){
             $form->validate( $post );
-            
+
             if( $form->isValid() ){
                 $item->name        = $post->name;
                 $item->description = $post->description;
                 $item->visible     = !is_null($post->visible);
                 $id = $item->put();
-                
+
                 if( $post->add_content ){
                     $content = Model_ItemContent::get();
                     $content->item_id = (null == $item->id ? $id : $item->id);
                     $content->put();
                 }
-                
-                foreach( $post->content as $key => $value ){
-                    $content = Model_ItemContent::get($key);
-                    if( isset($value['delete']) ){
-                        $content->delete();
-                    }
-                    else{
-                        $content->value = $value['value'];                    
-                        $content->put();                        
+
+                if( null !== $post->content ){
+                    foreach( $post->content as $key => $value ){
+                        $content = Model_ItemContent::get($key);
+                        if( isset($value['delete']) ){
+                            $content->delete();
+                        }
+                        else{
+                            $content->value = $value['value'];
+                            $content->put();
+                        }
                     }
                 }
-                
+
+
                 $this->_redirect( $this->getView()->Url(array(
                     'action'    => 'edit',
                     'id'        => (null == $item->id ? $id :$item->id )
                 )));
-                
+
             }
             else{
                 $this->getView()->errors = $form->getErrors();
             }
         }
-        
+
         $this->getView()->headScript()->append('/js/light-rte/jquery.rte.js');
         $this->getView()->headScript()->append(null, '$(function(){$(\'.rich-text-editor\').rte()})' );
         $this->getView()->Style()->append( '/js/light-rte/rte.css');
-        
+
         $this->getView()->content = $form;
         $this->getView()->actions = new Nano_Element('h2', null, $item->id ?
                         sprintf('Editing <em>%s</em>', $item->name ) : sprintf('New %s', $request->controller ));
-        
+
         $this->getView()->actions .= $this->getView()->Link('add ' . $request->controller
                                 , array('action' => 'add', 'id'=>null)
                                 , array('class'=>'button'));
-        
+
         $this->getView()->actions .= $this->getView()->Link(sprintf('list %ss', $request->controller)
                                 , array('action' => 'list', 'id'=>null)
                                 , array('class'=>'button'));
     }
 
+
     protected function deleteAction(){
         $request = $this->getRequest();
         $form    = new Nano_Form();
         $klass   = sprintf( 'Model_%s', ucfirst($request->getRouter()->controller));
-        
+
         if( is_numeric( $request->id ) ){
             $ids = array($request->id);
         }
         else{
             $ids = array_map( 'intval', explode(',', $request->id ));
         }
-        
+
         if( $request->isPost() && $post = $request->getPost() ){
             if( $post->confirm ){
                 foreach( $ids as $id ){
@@ -118,22 +122,22 @@ class Pico_AdminController extends Nano_Controller{
                 'id'    => null
             )));
         }
-    
+
         $list = new Nano_Element('ul');
-        
+
         foreach( $ids as $id ){
             $item = $klass::get($id, $klass);
             $list->addChild( 'li', null, $item->name );
         }
-        
+
         $form->addChild( 'p', null, sprintf(
             'You are about to delete the following %s'
             , count($ids) > 1 ? 'items' : 'item'
         ));
         $form->addChild( $list );
         $form->addChild( 'p', null, 'This cannot be undone, are you sure?');
-        
-    
+
+
         $form->addElements(array(
             'confirm' => array(
                 'prefix'    => '<p>',
@@ -148,7 +152,7 @@ class Pico_AdminController extends Nano_Controller{
                 'suffix'    => '</p>'
             )
         ));
-        
+
         $this->getView()->actions = '<h2>Confirm deletion</h2>';
         $this->getView()->content = $form;
     }
